+++ CodeStarterCamp_Week1/IcecreamMaker.swift
@@ -1,0 +1,25 @@
+//
+//  IcecreamMaker.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by JIWOONG on 2023/07/19.
+//
+//  Step 2: 아이스크림 그리기
+
+import Foundation
+
+func drawCream() {
+    let cream: String = "***********"
+    
+    for _ in 1...8 {
+        print(cream)
+    }
+}
+
+func drawBar() {
+    let bar: String = "    | |    "
+    
+    for _ in 1...4 {
+        print(bar)
+    }
+}
